Super sexy code refactoring

+++ rmpy/rmpy.py
@@ -3,32 +3,34 @@ All pretty pieces of this program is a result of Ballmer peak.
 Unfortunately the remaining 90% is the result of falling into the Ballmer tail.
 """
 
+import os
 import glob
 import shutil
 import logging
 import argparse
 import datetime
 
-from rmpy.auxiliary_tools import *
 from rmpy.config_tools import load_config
+from rmpy.auxiliary_tools import Codes, ask_confirmation, get_size, output, print_progress_bar
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Mine rm")
-    parser.add_argument("remove", nargs="*", help="File to delete")
+    parser = argparse.ArgumentParser()
+    parser.add_argument("remove", nargs="*", help="Delete files")
     parser.add_argument("-rec", dest="recover", help="Recover file")
     parser.add_argument("-re", dest="regex", help="Use regex")
     parser.add_argument("-e", dest="empty", action="store_true", help="Empty trash")
-    parser.add_argument("-s", dest="show", nargs="?", const=10, help="Show trash")
     parser.add_argument("-f", dest="force", action="store_true", help="Ignore nonexistent files")
+    parser.add_argument("-c", dest="confirm", action="store_true", help="Confirmation mode")
+    parser.add_argument("-s", dest="silent", action="store_true", help="Silent mode")
     parser.add_argument("--dry", action="store_true", help="Dry run")
-    parser.add_argument("--confirm", action="store_true", help="Confirm mode")
-    parser.add_argument("--silent", action="store_true", help="Silent mode")
-    parser.add_argument("--info", help="Info path")
-    parser.add_argument("--trash", help="Trash path")
     parser.add_argument("--log", help="Log path")
+    parser.add_argument("--show", nargs="?", const=10, help="Show trash")
+    parser.add_argument("--trash", help="Trash path")
     parser.add_argument("--config", help="Upload config")
     parser.add_argument("--policy", help="Recycle policy")
+    parser.add_argument("--size", help="Policy size")
+    parser.add_argument("--day", help="Policy day")
     args = parser.parse_args()
 
     config = load_config()
@@ -37,35 +39,43 @@ def main():
         config.update(load_config(new_config))
 
     dry = config["dry"]
-    silent = config["silent"]
-    confirm = config["confirm"]
-    policy = config["policy"]
+    day = config["day"]
+    size = config["size"]
     force = config["force"]
+    trash = config["trash"]
+    silent = config["silent"]
+    policy = config["policy"]
+    confirm = config["confirm"]
     log_path = config["log_path"]
-    info_path = config["info_path"]
-    trash_path = config["trash_path"]
 
     if args.dry:
         dry = True
+    if args.force:
+        force = True
     if args.silent:
         silent = True
     if args.confirm:
         confirm = True
-    if args.info:
-        info_path = args.info
+    if args.size:
+        size = int(args.size)
+    if args.day:
+        day = int(args.day)
     if args.trash:
-        trash_path = args.trash
+        trash = args.trash
     if args.log:
         log_path = args.log
     if args.policy:
         policy = args.policy
-    if args.force:
-        force = True
+
+    trash_path, info_path = os.path.join(trash, "files"), os.path.join(trash, "info")
+    if not os.path.exists(trash_path):
+        os.makedirs(trash_path)
+        os.makedirs(info_path)
 
     logging.basicConfig(format=u'%(levelname)-8s [%(asctime)s] %(message)s', level=logging.DEBUG, filename=log_path)
 
     if confirm:
-        confirmation()
+        ask_confirmation()
 
     if args.remove:
         files_num = 0
@@ -85,40 +95,53 @@ def main():
         output(silent, code, files_num, " File(s) were removed")
 
     if args.empty:
-        code = empty_trash(trash_path, info_path)
+        code = empty_trash(trash_path, info_path, dry, silent)
         output(silent, code, "Trash is empty now")
 
     if args.show:
-        show_trash(trash_path, num=int(args.show))
+        num = int(args.show)
+        trash_list = show_trash(trash_path, num)
+        output(silent, Codes.GOOD.value, "Last elements in trash: ", trash_list)
+        if not trash_list:
+            print("Trash is empty")
 
-    recycle_policy(policy, trash_path, info_path)
+    recycle_policy(policy, trash_path, info_path, day, size)
 
 
-def remove(target, trash_path, info_path, dry, silent, force=False):
+def remove(target, trash_path, info_path, dry=False, silent=False, force=False):
     files_num = 0
     code = Codes.GOOD.value
     target_path = os.path.abspath(target)
 
     try:
+        # New path for target and its info
         info = os.path.join(info_path, os.path.basename(target_path))
         trash = os.path.join(trash_path, os.path.basename(target_path))
+
+        # Looking for exception for dry-run case
         os.rename(target_path, target_path)
 
-        if not dry:
+        if dry:
+            if not silent:
+                print(target_path, " would be replaced to trash")
+
+        else:
+            # If target is a directory, then count the number of files inside
             if os.path.isdir(target_path):
                 files_num += sum([len(files) + len(dirs) for _, dirs, files in os.walk(target_path)])
 
-            # name conflict solution
+            # Name conflict solution
             if os.path.exists(trash):
                 count = 1
                 while True:
-                    if os.path.exists(trash + "_nfm_" + str(count)):
+                    if os.path.exists("{}_nfm_{}".format(trash, count)):
                         count += 1
                     else:
                         break
-                trash += "_nfm_" + str(count)
-                info += "_nfm_" + str(count)
+                trash += "_nfm_{}".format(count)
+                info += "_nfm_{}".format(count)
 
+            # Replace target to trash and create info file
             os.replace(target_path, trash)
             with open(info, 'w') as file:
                 file.write(target_path)
@@ -126,11 +149,8 @@ def remove(target, trash_path, info_path, dry, silent, force=False):
             files_num += 1
             logging.info(target_path + " File removed")
 
-        else:
-            if not silent:
-                print(target_path, " would be replaced to trash")
-
     except FileNotFoundError:
+        # Ignore exception for force mode case
         if not force:
             logging.error(target_path + " No such file to remove")
             code = Codes.NO_FILE.value
@@ -142,41 +162,47 @@ def remove(target, trash_path, info_path, dry, silent, force=False):
     return code, files_num
 
 
-def remove_by_regex(regex, trash_path, info_path, dry, silent, force=False):
-    files_num = 0
+def remove_by_regex(regex, trash_path, info_path, dry=False, silent=False, force=False):
+    files_removed = 0
     code = Codes.GOOD.value
+
+    # List of files names by regex
     files = glob.glob(regex)
 
     if files:
-        i = 0
-        l = len(files)
-        for obj in glob.glob(regex):
-            files_num += remove(obj, trash_path, info_path, dry, silent, force)[1]
-            i += 1
-            if not silent:
-                print_progress_bar(i, l, prefix='Progress:', suffix='Complete', length=50)
+        iteration = 0
+        total_files = len(files)
+
+        # Remove files and count their number
+        for file in files:
+            _, files_num = remove(file, trash_path, info_path, dry, silent, force)
+            files_removed += files_num
+
+            # Draw progress bar by current iteration
+            iteration += 1
+            if not dry and not silent:
+                print_progress_bar(iteration, total_files, prefix='Progress:', suffix='Complete', length=50)
 
     else:
         logging.info(regex + " No regex matches")
         code = Codes.NO_FILE.value
 
-    return code, files_num
+    return code, files_removed
 
 
 def recover_recursive(trash, path):
-    if not os.path.isdir(trash):
-        if os.path.exists(path):
-            if os.path.isdir(path):
-                shutil.rmtree(path)
-            os.replace(trash, path)
+    """Recursive function for solving conflicts, some code can be redundant"""
     if os.path.isdir(trash):
         if os.path.exists(path):
             if os.path.isdir(path):
                 for obj in os.listdir(trash):
+
                     trash_child = os.path.join(trash, obj)
                     path_child = os.path.join(path, obj)
+
                     if not os.path.isdir(trash_child):
                         os.replace(trash_child, path_child)
+
                     elif os.path.isdir(trash_child):
                         if os.path.exists(path_child):
                             recover_recursive(trash_child, path_child)
@@ -185,33 +211,46 @@ def recover_recursive(trash, path):
             else:
                 os.remove(path)
                 shutil.move(trash, path)
+    else:
+        if os.path.exists(path):
+            if os.path.isdir(path):
+                shutil.rmtree(path)
+            os.replace(trash, path)
 
 
-def recover(target, trash_path, info_path, dry, silent):
+def recover(target, trash_path, info_path, dry=False, silent=False):
     code = Codes.GOOD.value
+
     try:
+        # Paths to target and its info
         info = os.path.join(info_path, target)
         trash = os.path.join(trash_path, target)
 
+        # Reading previous file location
         with open(info, 'r') as file:
             path = file.read()
 
-        if not dry:
-            if not os.path.exists(os.path.dirname(path)):
-                os.makedirs(os.path.dirname(path))
-            os.replace(trash, path)
-            os.remove(info)
-            logging.info(target + " File recovered")
-
-        else:
+        if dry:
             if not silent:
                 print(path, " would be recovered")
+
+        else:
+            # Non-existent path conflict solution
+            if not os.path.exists(os.path.dirname(path)):
+                os.makedirs(os.path.dirname(path))
+
+            # Recover target and remove its info
+            os.replace(trash, path)
+            os.remove(info)
+
+            logging.info(target + " File recovered")
 
     except FileNotFoundError:
         logging.error(target + " No such file to recover")
         code = Codes.NO_FILE.value
 
     except OSError:
+        # Name conflict solution
         recover_recursive(trash, path)
         os.remove(info)
 
@@ -225,48 +264,51 @@ def recover(target, trash_path, info_path, dry, silent):
     return code
 
 
-def empty_trash(trash_path, info_path):
+def empty_trash(trash_path, info_path, dry=False, silent=False):
+    code = Codes.GOOD.value
+
     try:
-        for obj in os.listdir(trash_path):
-            trash = os.path.join(trash_path, obj)
+        if dry:
+            if not silent:
+                print("Trash gonna be empty")
 
-            if os.path.isfile(trash):
-                os.remove(trash)
-            elif os.path.isdir(trash):
-                shutil.rmtree(trash)
-            else:
-                os.unlink(trash)
+        else:
+            # Remove all files from trash
+            for file in os.listdir(trash_path):
+                trash = os.path.join(trash_path, file)
 
-        for obj in os.listdir(info_path):
-            os.remove(os.path.join(info_path, obj))
+                if os.path.isdir(trash):
+                    shutil.rmtree(trash)
+                else:
+                    os.remove(trash)
 
-        logging.info("Trash cleared")
-        return Codes.GOOD.value
+            # Remove all info files
+            for info in os.listdir(info_path):
+                os.remove(os.path.join(info_path, info))
+
+            logging.info("Trash cleared")
 
     except Exception:
         logging.error("Unknown error")
-        return Codes.BAD.value
+        code = Codes.BAD.value
+
+    return code
 
 
 def show_trash(trash_path, num):
     trash_list = os.listdir(trash_path)[-num:]
-    for file in trash_list:
-        print(file)
-    if trash_list:
-        print("Last %d elements in trash" % len(trash_list))
-    else:
-        print("Trash is empty")
 
-    return len(trash_list)
+    return trash_list
 
 
 def recycle_policy(policy, trash_path, info_path, day=6, size=100000):
     if policy == "time":
         if datetime.datetime.today().isoweekday() == day:
-            empty_trash(trash_path, info_path)
+            empty_trash(trash_path, info_path, dry=False, silent=True)
     if policy == "size":
         if get_size(trash_path) + get_size(info_path) >= size:
-            empty_trash(trash_path, info_path)
+            empty_trash(trash_path, info_path, dry=False, silent=True)
+
     return policy
